Extract thread inline payload helpers
